// src/api/__generated__/mining/model/actorLevelUpgradeMaterialResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Mining API
 * 挖矿服务接口
 * OpenAPI spec version: v1.0
 */
import type { ActorLevelUpgradeMaterialResponseDeploymentStatus } from './actorLevelUpgradeMaterialResponseDeploymentStatus';
import type { UpgradeRequirement } from './upgradeRequirement';

/**
 * 演员 NFT 咖位升级可用耗材
 */
export interface ActorLevelUpgradeMaterialResponse {
  /** 演员NFT tokenId，对应 actor_nft_position.token_id */
  actorId?: string;
  /** 演员合集ID */
  actorCollectionId?: string;
  /** 演员名称 */
  actorName?: string;
  /** 头像URL */
  avatarUrl?: string;
  /**
   * 等级
   * @minimum 1
   * @maximum 5
   */
  level?: number;
  /** 咖位名称 */
  levelName?: string;
  /**
   * 当前体力
   * @minimum 0
   * @maximum 168
   */
  stamina?: number;
  /** 体力上限 */
  staminaLimit?: number;
  /** 派遣状态 */
  deploymentStatus?: ActorLevelUpgradeMaterialResponseDeploymentStatus;
  upgradeRequirement?: UpgradeRequirement;
}